Optimized kernel reference fallbacks: MAX_POOL_2D et al
@tensorflow/micro

Remove conditional compilation for kernel tests for the following platforms: Xtensa
Add optimized kernel reference fallback code for the following platforms: Xtensa

bug=fixes Add reference fallbacks to optimized kernels: MAX_POOL_2D et al #1778
