+++ tensorflow/lite/micro/kernels/pooling_test.cc
@@ -1,4 +1,4 @@
-/* Copyright 2022 The TensorFlow Authors. All Rights Reserved.
+/* Copyright 2023 The TensorFlow Authors. All Rights Reserved.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -305,7 +305,6 @@ TF_LITE_MICRO_TEST(SimpleAveragePoolTestInt8PaddingSameStride1ActNone) {
       output_data);
 }
 
-#if !defined(XTENSA)
 TF_LITE_MICRO_TEST(SimpleAveragePoolTestInt16PaddingValidStride2ActNone) {
   int input_shape[] = {4, 1, 2, 4, 1};
   const int16_t input_values[] = {0, -24, 8, 16, 12, 8, -40, 28};
@@ -416,7 +415,6 @@ TF_LITE_MICRO_TEST(SimpleAveragePoolTestInt16PaddingSameStride1ActNone) {
       output_scale, output_zero_point, kTfLitePaddingValid, kTfLiteActNone,
       output_data);
 }
-#endif
 
 TF_LITE_MICRO_TEST(SimpleMaxPoolTestFloat) {
   int input_shape[] = {4, 1, 2, 4, 1};
@@ -616,7 +614,6 @@ TF_LITE_MICRO_TEST(MaxPoolTestInt8ActRelu6) {
       output_data);
 }
 
-#if !defined(XTENSA)
 TF_LITE_MICRO_TEST(SimpleMaxPoolTestInt16ActNone) {
   int input_shape[] = {4, 1, 2, 4, 1};
   const int16_t input_values1[] = {0, 6, 2, 4, 3, 2, 10, 7};
@@ -704,6 +701,5 @@ TF_LITE_MICRO_TEST(MaxPoolTestInt16ActRelu6) {
       output_scale, output_zero_point, kTfLitePaddingValid, kTfLiteActRelu6,
       output_data);
 }
-#endif
 
 TF_LITE_MICRO_TESTS_END
