+++ tensorflow/lite/micro/kernels/xtensa/pooling.cc
@@ -1,4 +1,4 @@
-/* Copyright 2022 The TensorFlow Authors. All Rights Reserved.
+/* Copyright 2023 The TensorFlow Authors. All Rights Reserved.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -12,13 +12,13 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
-#include "tensorflow/lite/kernels/internal/reference/pooling.h"
+
+#include "tensorflow/lite/micro/kernels/pooling.h"
 
 #include "tensorflow/lite/c/builtin_op_data.h"
 #include "tensorflow/lite/kernels/internal/tensor_ctypes.h"
 #include "tensorflow/lite/kernels/kernel_util.h"
 #include "tensorflow/lite/micro/kernels/kernel_util.h"
-#include "tensorflow/lite/micro/kernels/pooling.h"
 #include "tensorflow/lite/micro/kernels/xtensa/xtensa.h"
 #include "tensorflow/lite/micro/kernels/xtensa/xtensa_pooling.h"
 #include "tensorflow/lite/micro/micro_log.h"
@@ -27,206 +27,13 @@ namespace tflite {
 
 namespace {
 
-#if defined(HIFI5)
-
-struct OpData {
-  OpDataPooling reference_op_data;
-  int scratch_tensor_index;
-};
-
-TfLiteStatus AveragePrepareHifi(TfLiteContext* context, TfLiteNode* node) {
-  TF_LITE_ENSURE_STATUS(PoolingPrepare(context, node));
-  MicroContext* micro_context = GetMicroContext(context);
-  TfLiteTensor* input =
-      micro_context->AllocateTempInputTensor(node, kPoolingInputTensor);
-
-  if (input->type == kTfLiteInt8) {
-    const RuntimeShape& input_shape = GetTensorShape(input);
-    TfLiteTensor* output =
-        micro_context->AllocateTempInputTensor(node, kPoolingOutputTensor);
-    const RuntimeShape& output_shape = GetTensorShape(output);
-    micro_context->DeallocateTempTfLiteTensor(output);
-
-    const int depth = MatchingDim(input_shape, 3, output_shape, 3);
-    const int input_height = input_shape.Dims(1);
-    const int input_width = input_shape.Dims(2);
-    const int output_height = output_shape.Dims(1);
-    const int output_width = output_shape.Dims(2);
-
-    auto* params = reinterpret_cast<TfLitePoolParams*>(node->builtin_data);
-    auto* data = static_cast<OpData*>(node->user_data);
-
-    int required_scratch = xa_nn_avgpool_getsize(
-        depth, PREC_8, PREC_8, input_height, input_width, params->filter_height,
-        params->filter_width,
-        params->stride_width,                    // x_stride,
-        params->stride_height,                   // y_stride,
-        data->reference_op_data.padding.width,   // x_padding,
-        data->reference_op_data.padding.height,  // y_padding,
-        output_height, output_width, 0 /*NHWC input */, 0 /* NHWC output */);
-
-    if (required_scratch <= 0) {
-      MicroPrintf("Averagepool: xa_nn_avgpool_getsize failed");
-      return kTfLiteError;
-    }
-
-    TF_LITE_ENSURE_STATUS(context->RequestScratchBufferInArena(
-        context, required_scratch, &(data->scratch_tensor_index)));
-  }
-
-  micro_context->DeallocateTempTfLiteTensor(input);
-  return kTfLiteOk;
-}
-
-TfLiteStatus AverageEvalQuantizedHifi(TfLiteContext* context,
-                                      const TfLiteNode* node,
-                                      const TfLitePoolParams* params,
-                                      const OpData* data,
-                                      const TfLiteEvalTensor* input,
-                                      TfLiteEvalTensor* output) {
-  TFLITE_DCHECK(input->type == kTfLiteInt8);
-
-  const RuntimeShape& input_shape = tflite::micro::GetTensorShape(input);
-  const RuntimeShape& output_shape = tflite::micro::GetTensorShape(output);
-  const int batches = MatchingDim(input_shape, 0, output_shape, 0);
-  const int depth = MatchingDim(input_shape, 3, output_shape, 3);
-  const int input_height = input_shape.Dims(1);
-  const int input_width = input_shape.Dims(2);
-  const int output_height = output_shape.Dims(1);
-  const int output_width = output_shape.Dims(2);
-
-  void* p_scratch = static_cast<void*>(
-      context->GetScratchBuffer(context, data->scratch_tensor_index));
-
-  const int8_t* inp_data_ptr = tflite::micro::GetTensorData<int8_t>(input);
-  int8_t* out_data_ptr = tflite::micro::GetTensorData<int8_t>(output);
-
-  for (int batch = 0; batch < batches; ++batch) {
-    TF_LITE_ENSURE_EQ(
-        context,
-        xa_nn_avgpool_8(
-            &out_data_ptr[output_height * output_width * depth * batch],
-            const_cast<int8_t*>(
-                &inp_data_ptr[output_height * output_width * depth * batch]),
-            input_height, input_width, depth, params->filter_height,
-            params->filter_width, params->stride_width, params->stride_height,
-            data->reference_op_data.padding.width,
-            data->reference_op_data.padding.height, output_height, output_width,
-            0, 0, p_scratch),
-        0);
-  }
-
-  const int out_length = batches * output_height * output_width * depth;
-  TF_LITE_ENSURE_EQ(
-      context,
-      xa_nn_vec_activation_min_max_8_8(
-          out_data_ptr, out_data_ptr, data->reference_op_data.activation_min,
-          data->reference_op_data.activation_max, out_length),
-      0);
-
-  return kTfLiteOk;
-}
-
-TfLiteStatus MaxPrepareHifi(TfLiteContext* context, TfLiteNode* node) {
-  TF_LITE_ENSURE_STATUS(PoolingPrepare(context, node));
-
-  MicroContext* micro_context = GetMicroContext(context);
-
-  TfLiteTensor* input =
-      micro_context->AllocateTempInputTensor(node, kPoolingInputTensor);
-
-  if (input->type == kTfLiteInt8) {
-    auto* params = reinterpret_cast<TfLitePoolParams*>(node->builtin_data);
-    auto* data = static_cast<OpData*>(node->user_data);
-
-    const RuntimeShape& input_shape = GetTensorShape(input);
-    TfLiteTensor* output =
-        micro_context->AllocateTempOutputTensor(node, kPoolingOutputTensor);
-    const RuntimeShape& output_shape = GetTensorShape(output);
-    micro_context->DeallocateTempTfLiteTensor(output);
-
-    const int depth = MatchingDim(input_shape, 3, output_shape, 3);
-    const int input_height = input_shape.Dims(1);
-    const int input_width = input_shape.Dims(2);
-    const int output_height = output_shape.Dims(1);
-    const int output_width = output_shape.Dims(2);
-
-    int required_scratch = xa_nn_maxpool_getsize(
-        depth, PREC_8, PREC_8, input_height, input_width, params->filter_height,
-        params->filter_width,
-        params->stride_width,                    // x_stride,
-        params->stride_height,                   // y_stride,
-        data->reference_op_data.padding.width,   // x_padding,
-        data->reference_op_data.padding.height,  // y_padding,
-        output_height, output_width, 0 /* NHWC inpput */, 0 /* NHWC output */);
-
-    if (required_scratch <= 0) {
-      MicroPrintf("Maxpool: xa_nn_maxpool_getsize failed");
-      return kTfLiteError;
-    }
-
-    TF_LITE_ENSURE_STATUS(context->RequestScratchBufferInArena(
-        context, required_scratch, &(data->scratch_tensor_index)));
-  }
-
-  micro_context->DeallocateTempTfLiteTensor(input);
-  return kTfLiteOk;
-}
-
-TfLiteStatus MaxEvalQuantizedHifi(TfLiteContext* context, TfLiteNode* node,
-                                  TfLitePoolParams* params, const OpData* data,
-                                  const TfLiteEvalTensor* input,
-                                  TfLiteEvalTensor* output) {
-  const RuntimeShape& input_shape = tflite::micro::GetTensorShape(input);
-  const RuntimeShape& output_shape = tflite::micro::GetTensorShape(output);
-  const int batches = MatchingDim(input_shape, 0, output_shape, 0);
-  const int depth = MatchingDim(input_shape, 3, output_shape, 3);
-  const int input_height = input_shape.Dims(1);
-  const int input_width = input_shape.Dims(2);
-  const int output_height = output_shape.Dims(1);
-  const int output_width = output_shape.Dims(2);
-
-  void* p_scratch = static_cast<void*>(
-      context->GetScratchBuffer(context, data->scratch_tensor_index));
-
-  const int8_t* inp_data_ptr = tflite::micro::GetTensorData<int8_t>(input);
-  int8_t* out_data_ptr = tflite::micro::GetTensorData<int8_t>(output);
-
-  for (int batch = 0; batch < batches; ++batch) {
-    TF_LITE_ENSURE_EQ(
-        context,
-        xa_nn_maxpool_8(
-            &out_data_ptr[output_height * output_width * depth * batch],
-            const_cast<int8_t*>(
-                &inp_data_ptr[output_height * output_width * depth * batch]),
-            input_height, input_width, depth, params->filter_height,
-            params->filter_width, params->stride_width, params->stride_height,
-            data->reference_op_data.padding.width,
-            data->reference_op_data.padding.height, output_height, output_width,
-            0, 0, p_scratch),
-        0);
-  }
-
-  const int out_length = batches * output_height * output_width * depth;
-  TF_LITE_ENSURE_EQ(
-      context,
-      xa_nn_vec_activation_min_max_8_8(
-          out_data_ptr, out_data_ptr, data->reference_op_data.activation_min,
-          data->reference_op_data.activation_max, out_length),
-      0);
-
-  return kTfLiteOk;
-}
-
-#endif  // defined(HIFI5)
-
 TfLiteStatus AverageEval(TfLiteContext* context, TfLiteNode* node) {
   TFLITE_DCHECK(node->builtin_data != nullptr);
   auto* params = reinterpret_cast<TfLitePoolParams*>(node->builtin_data);
 
   TFLITE_DCHECK(node->user_data != nullptr);
 #if defined(HIFI5)
-  const OpData* op_data = static_cast<const OpData*>(node->user_data);
+  auto* op_data = static_cast<const XtensaOpDataPooling*>(node->user_data);
   const OpDataPooling* reference_op_data = &(op_data->reference_op_data);
 #else
   const OpDataPooling* reference_op_data =
@@ -243,7 +50,8 @@ TfLiteStatus AverageEval(TfLiteContext* context, TfLiteNode* node) {
     case kTfLiteFloat32: {
       AveragePoolingEvalFloat(context, node, params, reference_op_data, input,
                               output);
-    } break;
+      break;
+    }
     case kTfLiteInt8: {
 #if defined(HIFI5)
       AverageEvalQuantizedHifi(context, node, params, op_data, input, output);
@@ -255,11 +63,18 @@ TfLiteStatus AverageEval(TfLiteContext* context, TfLiteNode* node) {
       AveragePoolingEvalQuantized<int8_t>(context, node, params,
                                           reference_op_data, input, output);
 #endif
-    } break;
-    default:
+      break;
+    }
+    case kTfLiteInt16: {
+      AveragePoolingEvalQuantized<int16_t>(context, node, params,
+                                           reference_op_data, input, output);
+      break;
+    }
+    default: {
       MicroPrintf("Input type %s is not currently supported",
                   TfLiteTypeGetName(input->type));
       return kTfLiteError;
+    }
   }
   return kTfLiteOk;
 }
@@ -270,7 +85,7 @@ TfLiteStatus MaxEval(TfLiteContext* context, TfLiteNode* node) {
 
   TFLITE_DCHECK(node->user_data != nullptr);
 #if defined(HIFI5)
-  const OpData* op_data = static_cast<const OpData*>(node->user_data);
+  auto* op_data = static_cast<const XtensaOpDataPooling*>(node->user_data);
   const OpDataPooling* reference_op_data = &(op_data->reference_op_data);
 #else
   const OpDataPooling* reference_op_data =
@@ -286,7 +101,8 @@ TfLiteStatus MaxEval(TfLiteContext* context, TfLiteNode* node) {
     case kTfLiteFloat32: {
       MaxPoolingEvalFloat(context, node, params, reference_op_data, input,
                           output);
-    } break;
+      break;
+    }
     case kTfLiteInt8: {
 #if defined(HIFI5)
       MaxEvalQuantizedHifi(context, node, params, op_data, input, output);
@@ -298,50 +114,54 @@ TfLiteStatus MaxEval(TfLiteContext* context, TfLiteNode* node) {
       MaxPoolingEvalQuantized<int8_t>(context, node, params, reference_op_data,
                                       input, output);
 #endif
-    } break;
-    default:
+      break;
+    }
+    case kTfLiteInt16: {
+      MaxPoolingEvalQuantized<int16_t>(context, node, params, reference_op_data,
+                                       input, output);
+      break;
+    }
+    default: {
       MicroPrintf("Type %s not currently supported.",
                   TfLiteTypeGetName(input->type));
       return kTfLiteError;
+    }
   }
   return kTfLiteOk;
-}
-
-void* Init(TfLiteContext* context, const char* buffer, size_t length) {
-  TFLITE_DCHECK(context->AllocatePersistentBuffer != nullptr);
-#if defined(HIFI5)
-  return context->AllocatePersistentBuffer(context, sizeof(OpData));
-#elif defined(VISION_P6)
-  if (InitXtensaContext()) {
-    return nullptr;
-  }
-  return context->AllocatePersistentBuffer(context,
-                                           sizeof(XtensaOpDataPooling));
-#else
-  return context->AllocatePersistentBuffer(context, sizeof(OpDataPooling));
-#endif
 }
 
 }  // namespace
 
 TfLiteRegistration Register_AVERAGE_POOL_2D() {
 #if defined(HIFI5)
-  return tflite::micro::RegisterOp(Init, AveragePrepareHifi, AverageEval);
+  return tflite::micro::RegisterOp(XtensaPoolingInit, AveragePrepareHifi,
+                                   AverageEval);
 #elif defined(VISION_P6)
-  return tflite::micro::RegisterOp(Init, AvgPoolingPrepareVision, AverageEval);
+  return tflite::micro::RegisterOp(XtensaPoolingInit, AvgPoolingPrepareVision,
+                                   AverageEval);
 #else
-  return tflite::micro::RegisterOp(Init, PoolingPrepare, AverageEval);
+  return tflite::micro::RegisterOp(XtensaPoolingInit, PoolingPrepare,
+                                   AverageEval);
 #endif
 }
 
 TfLiteRegistration Register_MAX_POOL_2D() {
 #if defined(HIFI5)
-  return tflite::micro::RegisterOp(Init, MaxPrepareHifi, MaxEval);
+  return tflite::micro::RegisterOp(XtensaPoolingInit, MaxPrepareHifi, MaxEval);
 #elif defined(VISION_P6)
-  return tflite::micro::RegisterOp(Init, MaxPoolingPrepareVision, MaxEval);
+  return tflite::micro::RegisterOp(XtensaPoolingInit, MaxPoolingPrepareVision,
+                                   MaxEval);
 #else
-  return tflite::micro::RegisterOp(Init, PoolingPrepare, MaxEval);
+  return tflite::micro::RegisterOp(XtensaPoolingInit, PoolingPrepare, MaxEval);
 #endif
+}
+
+TfLiteRegistration Register_AVERAGE_POOL_2D_INT16() {
+  return Register_AVERAGE_POOL_2D();
+}
+
+TfLiteRegistration Register_MAX_POOL_2D_INT16() {
+  return Register_MAX_POOL_2D();
 }
 
 }  // namespace tflite
