+++ tensorflow/lite/micro/kernels/xtensa/pooling_vision.cc
@@ -1,4 +1,4 @@
-/* Copyright 2022 The TensorFlow Authors. All Rights Reserved.
+/* Copyright 2023 The TensorFlow Authors. All Rights Reserved.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -46,41 +46,43 @@ TfLiteStatus PoolingPrepareVision(TfLiteContext* context, TfLiteNode* node,
       micro_context->AllocateTempInputTensor(node, kPoolingInputTensor);
   TF_LITE_ENSURE(context, input != nullptr);
 
-  uint32_t context_size = 0;
-  uint32_t status = xiPoolGetMemReqd_Context(&context_size);
-  TFLITE_DCHECK(status == 0);
-  if (context_size) {
-    void* context_data =
-        context->AllocatePersistentBuffer(context, context_size);
-    if (context_data == nullptr) {
+  if (input->type == kTfLiteInt8) {
+    uint32_t context_size = 0;
+    uint32_t status = xiPoolGetMemReqd_Context(&context_size);
+    TFLITE_DCHECK(status == 0);
+    if (context_size) {
+      void* context_data =
+          context->AllocatePersistentBuffer(context, context_size);
+      if (context_data == nullptr) {
+        return kTfLiteError;
+      }
+      data->p_context = reinterpret_cast<uint8_t*>(context_data);
+      data->context_size = context_size;
+    }
+
+    uint32_t input_dims[4] = {1, 1, 1, 1};
+    uint32_t output_dims[4] = {1, 1, 1, 1};
+    for (int i = 0; i < NumDimensions(input); i++) {
+      input_dims[i] =
+          std::max(1, SizeOfDimension(input, NumDimensions(input) - 1 - i));
+    }
+    for (int i = 0; i < NumDimensions(output); i++) {
+      output_dims[i] =
+          std::max(1, SizeOfDimension(output, NumDimensions(output) - 1 - i));
+    }
+
+    status = xiPoolSetContext(
+        data->p_context, data->context_size, input_dims[0], input_dims[1],
+        input_dims[2], input_dims[3], output_dims[0], output_dims[1],
+        output_dims[2], params.filter_width, params.filter_height,
+        params.stride_width, params.stride_height,
+        data->reference_op_data.padding.width,
+        data->reference_op_data.padding.height, input->params.zero_point,
+        output->params.zero_point, data->reference_op_data.activation_min,
+        data->reference_op_data.activation_max, pool_type);
+    if (status) {
       return kTfLiteError;
     }
-    data->p_context = reinterpret_cast<uint8_t*>(context_data);
-    data->context_size = context_size;
-  }
-
-  uint32_t input_dims[4] = {1, 1, 1, 1};
-  uint32_t output_dims[4] = {1, 1, 1, 1};
-  for (int i = 0; i < NumDimensions(input); i++) {
-    input_dims[i] =
-        std::max(1, SizeOfDimension(input, NumDimensions(input) - 1 - i));
-  }
-  for (int i = 0; i < NumDimensions(output); i++) {
-    output_dims[i] =
-        std::max(1, SizeOfDimension(output, NumDimensions(output) - 1 - i));
-  }
-
-  status = xiPoolSetContext(
-      data->p_context, data->context_size, input_dims[0], input_dims[1],
-      input_dims[2], input_dims[3], output_dims[0], output_dims[1],
-      output_dims[2], params.filter_width, params.filter_height,
-      params.stride_width, params.stride_height,
-      data->reference_op_data.padding.width,
-      data->reference_op_data.padding.height, input->params.zero_point,
-      output->params.zero_point, data->reference_op_data.activation_min,
-      data->reference_op_data.activation_max, pool_type);
-  if (status) {
-    return kTfLiteError;
   }
 
   micro_context->DeallocateTempTfLiteTensor(output);
